tests/wpt/tests/css/css-sizing/animation/min-height-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>min-height interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-min-height">
<meta name="assert" content="min-height supports animation by computed value">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  min-height: 30px;
}
.target {
  width: 10px;
  height: 0px;
  background-color: black;
  display: inline-block;
  min-height: 10px;
}
.expected {
  background-color: green;
}
</style>

<body></body>

<script>
test_interpolation({
  property: 'min-height',
  from: neutralKeyframe,
  to: '20px',
}, [
  {at: -0.5, expect: '5px'},
  {at: 0, expect: '10px'},
  {at: 0.3, expect: '13px'},
  {at: 0.6, expect: '16px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '25px'},
]);

test_no_interpolation({
  property: 'min-height',
  from: 'initial',
  to: '20px',
});

test_interpolation({
  property: 'min-height',
  from: 'inherit',
  to: '20px',
}, [
  {at: -0.5, expect: '35px'},
  {at: 0, expect: '30px'},
  {at: 0.3, expect: '27px'},
  {at: 0.6, expect: '24px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '15px'},
]);

test_no_interpolation({
  property: 'min-height',
  from: 'unset',
  to: '20px',
});

test_no_interpolation({
  property: 'min-height',
  from: 'auto',
  to: '20px',
});

test_interpolation({
  property: 'min-height',
  from: '0px',
  to: '100px',
}, [
  {at: -0.5, expect: '0'}, // CSS min-height can't be negative.
  {at: 0, expect: '0'},
  {at: 0.3, expect: '30px'},
  {at: 0.6, expect: '60px'},
  {at: 1, expect: '100px'},
  {at: 1.5, expect: '150px'}
]);

test_no_interpolation({
  property: 'min-height',
  from: 'auto',
  to: '0px',
});

test_no_interpolation({
  property: 'min-height',
  from: 'fit-content',
  to: 'min-content',
});

test_no_interpolation({
  property: 'min-height',
  from: 'max-content',
  to: 'stretch',
});

test_no_interpolation({
  property: 'min-height',
  from: 'fit-content',
  to: neutralKeyframe,
});

test_no_interpolation({
  property: 'min-height',
  from: neutralKeyframe,
  to: 'min-content',
});

</script>
