third_party/blink/web_tests/external/wpt/css/css-sizing/animation/min-height-interpolation-expected.txt

This is a testharness.js-based test.
Found 42 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (-0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.3) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (-0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.3) should be [max-content]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (0.6) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (1) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-height> from [max-content] to [stretch] at (1.5) should be [stretch]
  assert_true: 'to' value should be supported expected true got false
Harness: the test ran to completion.
